add app layout optimization
