+++ src/app/layout.tsx
@@ -7,8 +7,8 @@ import { UtilityProvider } from "@/context/UtilityContext";
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "Wordle",
+  description: "Wordle Game",
 };
 
 export default function RootLayout({
